+++ src/main/java/com/zero/cohousesever/post/dto/post/PostResponse.java
@@ -1,4 +1,4 @@
-package com.zero.cohousesever.post.dto;
+package com.zero.cohousesever.post.dto.post;
 
 import com.zero.cohousesever.post.entity.Post;
 import com.zero.cohousesever.post.type.PostType;
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class PostResponse {
