+++ src/main/java/com/zero/cohousesever/post/dto/postLike/PostLikeCountResponse.java
@@ -1,9 +1,8 @@
-package com.zero.cohousesever.post.dto;
+package com.zero.cohousesever.post.dto.postLike;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 /**
  * 특정 게시글의 좋아요 개수만 응답.
