+++ src/main/java/com/zero/cohousesever/post/dto/postLike/PostLikeStatusResponse.java
@@ -1,4 +1,4 @@
-package com.zero.cohousesever.post.dto;
+package com.zero.cohousesever.post.dto.postLike;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,7 +10,6 @@ import lombok.NoArgsConstructor;
  * -특정 게시글에 대해 현재 사용자가 좋아요를 눌렀는지 여부 응답.
  */
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class PostLikeStatusResponse {
